caption-picture/requirements.txt
transformers
gradio
pillow
